Move CustomNotification component inside story
