--- src/Notification/story/CustomNotification.module.css
@@ -1,16 +1,0 @@
-.root {
-  &:global {
-    &.default {
-      background-color: lightblue;
-    }
-    &.success {
-      background-color: lightgreen;
-    }
-    &.warning {
-      background-color: lightsalmon;
-    }
-    &.error {
-      background-color: lightcoral;
-    }
-  }
-}
